refactor: 유효성 검증 함수 분리
